Update Cel.py
Clearly, I changed a big part of the code. The structure mostly, then if any error occurs or there're changes in the parameters used in a certain product one is able to fixed easier instead of getting lost in a dense jungle of an Script.  Also part of the movements of the mouse with pyaitogui, I replaced the movements to press the button with directly pressing the keys Ctrl + P, that way any monitor is able to use the Script.